Added IO throttling for the merger pipeline.
We rely on async-speed-limit for the throttling.
The throttling of the "split download" will be added in
a separate PR.

This PR also does some minor refactoring on the merge pipeline
parameters and introduce two counters with the indexer write io,
and the merge executor write io.

Relying on async-speed-limit